test(core): consolidate translation fallback e2e tests into facet.e2e-spec.ts
Move the 'Facet translation language fallback' describe block from the
separate facet-translation-fallback.e2e-spec.ts into the existing
facet.e2e-spec.ts, as suggested in code review.
